refactor: extract seven eleven code commonly used
